[REFACTOR] [Setting] 단위 별로 파일 리팩토링
